Extract ProjectsController
